Revert "Make User.active default to true."
This reverts commit ccda1b39df3f1b90597478db5c46e7d1ac5541c0.

Well, this didn't work. ¯\_(ツ)_/¯

So, now there's an imcompatibility if you're using a current version of Meteroid
with an instance of Mete before fe561e1 (i.e. not updated for about a year).
To be clear: This will not mess up any data.
But all accounts will be displayed as inactive and editing or creating new
or existing users won't be possible until the Mete instance is upgraded.